fix dtype of index for empty edge_labels
super artificial example, but it's my testdata in ilastik - super-pixel
edges are only in y-direction. So for x, the dense_edge_tables will be
empty. Newer version of pandas seem to infer a default index dtype after
of the index of the dataframe after merge 'object', which made difficulties
when serializing.
I think it makes sense to always get the same dtype for the index.
Note: I checked and couldn't even find what the last pandas version is
that would support uint32 index. It will be pd.In64Index for any kind of
integers, so I removed the misleading naming and left a comment in the code.